# Conflicts:
#	app/src/main/java/example/Calculator.java

+++ app/src/main/java/example/Calculator.java
@@ -2,22 +2,22 @@ package example;
 
 public class Calculator {
     public int add(int a, int b) {
-
-        return a + b ;
+        // TODO: Implement this method
+        return a+b;
     }
 
     public int minus(int a, int b) {
         // TODO: Implement this method
-        return a - b;
+        return a-b;
     }
 
     public int multiply(int a, int b) {
         // TODO: Implement this method
-        return a *b ;
+        return a*b;
     }
 
     public double divide(int a, int b) {
         // TODO: Implement this method
-        return a/b;
+        return (double) a /b;
     }
 }
